+++ src/transformer/descriptor/helper/helper.ts
@@ -15,24 +15,26 @@ export namespace TypescriptHelper {
     export function GetDeclarationFromNode(node: ts.Node): ts.Declaration {
         const typeChecker: ts.TypeChecker = TypeChecker();
         const symbol: ts.Symbol = typeChecker.getSymbolAtLocation(node);
+
+        return GetDeclarationFromSymbol(symbol);
+    }
+
+    export function GetDeclarationFromSymbol(symbol: ts.Symbol): ts.Declaration {
         const declaration: ts.Declaration = GetFirstValidDeclaration(symbol.declarations);
 
-        if (ts.isImportSpecifier(declaration)) {
-            return GetDeclarationForImport(declaration) as ts.Declaration;
-        }
-
-        if (ts.isImportEqualsDeclaration(declaration)) {
-            return GetDeclarationFromNode(declaration.moduleReference);
+        if (ts.isImportSpecifier(declaration) || ts.isImportEqualsDeclaration(declaration)) {
+            return GetDeclarationForImport(declaration);
         }
 
         return declaration;
     }
 
-    export function GetDeclarationForImport(node: ts.ImportClause | ts.ImportSpecifier): ts.TypeNode | ts.Declaration {
+    export function GetDeclarationForImport(node: ts.ImportClause | ts.ImportSpecifier | ts.ImportEqualsDeclaration): ts.Declaration {
         const typeChecker: ts.TypeChecker = TypeChecker();
         const symbol: ts.Symbol = typeChecker.getSymbolAtLocation(node.name);
-        const declaredType: ts.Type = typeChecker.getDeclaredTypeOfSymbol(symbol);
-        return GetDeclarationFromType(declaredType);
+        const originalSymbol: ts.Symbol = typeChecker.getAliasedSymbol(symbol);
+
+        return GetFirstValidDeclaration(originalSymbol.declarations);
     }
 
     export function GetParameterOfNode(node: ts.EntityName): ts.NodeArray<ts.TypeParameterDeclaration> {
@@ -41,21 +43,11 @@ export namespace TypescriptHelper {
         return (declaration as Declaration).typeParameters;
     }
 
-    export function GetDeclarationFromType(type: ts.Type): ts.TypeNode | ts.Declaration {
-        if (type.symbol && type.symbol.declarations) {
-            return GetFirstValidDeclaration(type.symbol.declarations);
-        } else if (type.aliasSymbol && type.aliasSymbol.declarations) {
-            return GetFirstValidDeclaration(type.aliasSymbol.declarations);
-        }
-
-        return TypeChecker().typeToTypeNode(type);
-    }
-
     export function GetTypeParameterOwnerMock(declaration: ts.Declaration): ts.Declaration {
         const typeDeclaration: ts.Declaration = ts.getTypeParameterOwner(declaration);
 
         // THIS IS TO FIX A MISSING IMPLEMENTATION IN TYPESCRIPT https://github.com/microsoft/TypeScript/blob/ba5e86f1406f39e89d56d4b32fd6ff8de09a0bf3/src/compiler/utilities.ts#L5138
-        if ((typeDeclaration as Declaration).typeParameters) {
+        if (typeDeclaration && (typeDeclaration as Declaration).typeParameters) {
             return typeDeclaration;
         }
 
@@ -66,9 +58,18 @@ export namespace TypescriptHelper {
         }
     }
 
+    export function GetStringPropertyName(propertyName: ts.PropertyName): string {
+        if (!ts.isComputedPropertyName(propertyName)) {
+            return propertyName.text;
+        }
+
+        const symbol: ts.Symbol = TypeChecker().getSymbolAtLocation(propertyName);
+        return symbol.escapedName.toString();
+    }
+
     function GetFirstValidDeclaration(declarations: ts.Declaration[]): ts.Declaration {
         return declarations.find((declaration: ts.Declaration) => {
             return !ts.isVariableDeclaration(declaration);
-        });
+        }) || declarations[0];
     }
 }