+++ src/transformer/descriptor/mock/mockDeclarationName.ts
@@ -1,12 +1,13 @@
 import * as ts from 'typescript';
 
-const VARIABLE_DECLARATION_PREFIX: string = '_s';
-const VARIABLE_PARAMETER_SET_PREFIX: string = '_';
-
-export function GetMockDeclarationName(name: ts.Identifier): ts.Identifier {
-    return ts.createIdentifier(VARIABLE_DECLARATION_PREFIX + name.escapedText);
+export function GetMockInternalValuesName(): ts.Identifier {
+    return ts.createIdentifier('declarations');
 }
 
-export function GetMockSetParameterName(name: ts.Identifier): ts.Identifier {
-    return ts.createIdentifier(VARIABLE_PARAMETER_SET_PREFIX + name.escapedText);
+export function GetMockObjectReturnValueName(): ts.Identifier {
+    return ts.createIdentifier('mockObjectReturnValue');
+}
+
+export function GetMockSetParameterName(): ts.Identifier {
+    return ts.createIdentifier('value');
 }