+++ src/transformer/descriptor/mock/mockProperties.ts
@@ -3,19 +3,21 @@ import { Scope } from '../../scope/scope';
 import { GetDescriptor } from '../descriptor';
 import { IsTypescriptType } from '../tsLibs/typecriptLibs';
 import { GetMockCall } from './mockCall';
-import { GetMockDeclarationName } from './mockDeclarationName';
 import { GetMockProperty } from './mockProperty';
 
 export function GetMockPropertiesFromSymbol(propertiesSymbol: ts.Symbol[], signatures: ReadonlyArray<ts.Signature>, scope: Scope): ts.Expression {
     const properties: ts.Declaration[] = propertiesSymbol.map((prop: ts.Symbol) => {
         return prop.declarations[0];
     });
+    const signaturesDeclarations: ts.Declaration[] = signatures.map((signature: ts.Signature) => {
+        return signature.declaration;
+    });
 
-    return GetMockPropertiesFromDeclarations(properties, signatures, scope);
+    return GetMockPropertiesFromDeclarations(properties, signaturesDeclarations, scope);
 }
 
-export function GetMockPropertiesFromDeclarations(list: ts.Declaration[], signatures: ReadonlyArray<ts.Signature>, scope: Scope): ts.CallExpression {
-    const propertiesFilter: ts.Declaration[] = list.filter((member: ts.PropertyDeclaration) => {
+export function GetMockPropertiesFromDeclarations(list: ReadonlyArray<ts.Declaration>, signatures: ReadonlyArray<ts.Declaration>, scope: Scope): ts.CallExpression {
+    const propertiesFilter: ts.Declaration[] = list.filter((member: ts.PropertySignature) => {
         const hasModifiers: boolean = !!member.modifiers;
 
         if (IsTypescriptType(member)) { // This is a current workaround to safe fail extends of TypescriptLibs
@@ -31,17 +33,12 @@ export function GetMockPropertiesFromDeclarations(list: ts.Declaration[], signat
         }).length === 0;
     });
 
-    const variableDeclarations: ts.VariableDeclaration[] = propertiesFilter.map((member: ts.PropertySignature) => {
-        const name: ts.Identifier = GetMockDeclarationName(member.name as ts.Identifier);
-        return ts.createVariableDeclaration(name);
-    });
-
     const accessorDeclaration: ts.PropertyAssignment[] = propertiesFilter.map(
         (member: ts.PropertySignature): ts.PropertyAssignment => {
             return GetMockProperty(member, scope);
         },
     );
 
-    const signaturesDescriptor: ts.Expression = signatures.length > 0 ? GetDescriptor(signatures[0].declaration, scope) : null;
-    return GetMockCall(variableDeclarations, accessorDeclaration, signaturesDescriptor);
+    const signaturesDescriptor: ts.Expression = signatures.length > 0 ? GetDescriptor(signatures[0], scope) : null;
+    return GetMockCall(accessorDeclaration, signaturesDescriptor);
 }