+++ src/transformer/mockDefiner/mockDefiner.ts
@@ -2,6 +2,7 @@ import * as ts from 'typescript';
 import { GetTsAutoMockCacheOptions, TsAutoMockCacheOptions } from '../../options/cache';
 import { GetDescriptor } from '../descriptor/descriptor';
 import { GetProperties } from '../descriptor/properties/properties';
+import { GetTypeofEnumDescriptor } from '../descriptor/typeQuery/enumTypeQuery';
 import { TypescriptCreator } from '../helper/creator';
 import { createImportOnIdentifier } from '../helper/import';
 import { MockGenericParameter } from '../mockGeneric/mockGenericParameter';
@@ -12,9 +13,6 @@ import { FactoryUniqueName, PossibleDeclaration } from './factoryUniqueName';
 import { ModuleName } from './modules/moduleName';
 import { ModuleNameIdentifier } from './modules/moduleNameIdentifier';
 import { ModulesImportUrl } from './modules/modulesImportUrl';
-
-// tslint:disable-next-line:no-any
-const urlSlug: any = require('url-slug');
 
 interface FactoryRegistrationPerFile {
     [key: string]: Array<{
@@ -32,9 +30,12 @@ interface FactoryIntersectionRegistrationPerFile {
 
 export class MockDefiner {
     private _neededImportIdentifierPerFile: { [key: string]: Array<ModuleNameIdentifier> } = {};
+    private _internalModuleImportIdentifierPerFile: { [key: string]: { [key in ModuleName]: ts.Identifier } } = {};
     private _factoryRegistrationsPerFile: FactoryRegistrationPerFile = {};
+    private _registerMockFactoryRegistrationsPerFile: FactoryRegistrationPerFile = {};
     private _factoryIntersectionsRegistrationsPerFile: FactoryIntersectionRegistrationPerFile = {};
     private _factoryCache: DeclarationCache;
+    private _registerMockFactoryCache: DeclarationCache;
     private _declarationCache: DeclarationCache;
     private _factoryIntersectionCache: DeclarationListCache;
     private _fileName: string;
@@ -46,6 +47,7 @@ export class MockDefiner {
         this._declarationCache = new DeclarationCache();
         this._factoryIntersectionCache = new DeclarationListCache();
         this._factoryUniqueName = new FactoryUniqueName();
+        this._registerMockFactoryCache = new DeclarationCache();
         this._cacheEnabled = GetTsAutoMockCacheOptions();
     }
 
@@ -62,15 +64,24 @@ export class MockDefiner {
     }
 
     public setTsAutoMockImportIdentifier(): void {
-        if (!this._neededImportIdentifierPerFile[this._fileName]) {
-            this._neededImportIdentifierPerFile[this._fileName] = Object.keys(ModulesImportUrl).map((key: ModuleName) => {
-                return {
-                    name: key,
-                    moduleUrl: ModulesImportUrl[key],
-                    identifier: this._createUniqueFileName(key),
-                };
-            });
+        if (this._internalModuleImportIdentifierPerFile[this._fileName]) {
+            return;
         }
+
+        this._internalModuleImportIdentifierPerFile[this._fileName] = {
+            [ModuleName.Extension]: ts.createFileLevelUniqueName(ModuleName.Extension),
+            [ModuleName.Merge]: ts.createFileLevelUniqueName(ModuleName.Merge),
+            [ModuleName.Repository]: ts.createFileLevelUniqueName(ModuleName.Repository),
+        };
+
+        this._neededImportIdentifierPerFile[this._fileName] = this._neededImportIdentifierPerFile[this._fileName] || [];
+
+        Array.prototype.push.apply(this._neededImportIdentifierPerFile[this._fileName], Object.keys(ModulesImportUrl).map((key: ModuleName) => {
+            return {
+                moduleUrl: ModulesImportUrl[key],
+                identifier: this._internalModuleImportIdentifierPerFile[this._fileName][key],
+            };
+        }));
     }
 
     public getCurrentModuleIdentifier(module: ModuleName): ts.Identifier {
@@ -82,6 +93,7 @@ export class MockDefiner {
             ...this._getImportsToAddInFile(sourceFile),
             ...this._getExportsToAddInFile(sourceFile),
             ...this._getExportsIntersectionToAddInFile(sourceFile),
+            ...this._getRegisterMockInFile(sourceFile),
         ];
     }
 
@@ -93,10 +105,32 @@ export class MockDefiner {
         }
         this._factoryRegistrationsPerFile[sourceFile.fileName] = [];
         this._factoryIntersectionsRegistrationsPerFile[sourceFile.fileName] = [];
+        this._registerMockFactoryRegistrationsPerFile[sourceFile.fileName] = [];
     }
 
-    public getMockFactory(declaration: ts.Declaration): ts.Expression {
-        const key: string = this._getMockFactoryId(declaration);
+    public createMockFactory(declaration: ts.Declaration): void {
+        const thisFileName: string = this._fileName;
+
+        const key: string = this.getDeclarationKeyMap(declaration);
+
+        this._factoryCache.set(declaration, key);
+
+        this._factoryRegistrationsPerFile[thisFileName] = this._factoryRegistrationsPerFile[thisFileName] || [];
+
+        const descriptor: ts.Expression = GetDescriptor(declaration, new Scope(key));
+
+        const mockGenericParameter: ts.ParameterDeclaration = this._getMockGenericParameter();
+
+        const factory: ts.FunctionExpression = TypescriptCreator.createFunctionExpressionReturn(descriptor, [mockGenericParameter]);
+
+        this._factoryRegistrationsPerFile[thisFileName].push({
+            key: declaration,
+            factory,
+        });
+    }
+
+    public getMockFactoryTypeofEnum(declaration: ts.EnumDeclaration): ts.Expression {
+        const key: string = this._getMockFactoryIdForTypeofEnum(declaration);
 
         return this.getMockFactoryByKey(key);
     }
@@ -123,8 +157,20 @@ export class MockDefiner {
         return this._declarationCache.get(declaration);
     }
 
-    private _createUniqueFileName(name: string): ts.Identifier {
-        return ts.createFileLevelUniqueName(`${urlSlug(this._fileName, '_')}_${name}`);
+    public storeRegisterMockFor(declaration: ts.Declaration, factory: ts.FunctionExpression): void {
+        const key: string = this.getDeclarationKeyMap(declaration);
+
+        this._registerMockFactoryCache.set(declaration, key);
+
+        this._registerMockFactoryRegistrationsPerFile[this._fileName] = this._registerMockFactoryRegistrationsPerFile[this._fileName] || [];
+        this._registerMockFactoryRegistrationsPerFile[this._fileName].push({
+            key: declaration,
+            factory,
+        });
+    }
+
+    public hasMockForDeclaration(declaration: ts.Declaration): boolean {
+        return this._factoryCache.has(declaration) || this._registerMockFactoryCache.has(declaration);
     }
 
     private _mockRepositoryAccess(filename: string): ts.Expression {
@@ -140,29 +186,22 @@ export class MockDefiner {
     }
 
     private _getModuleIdentifier(fileName: string, module: ModuleName): ts.Identifier {
-        return this._neededImportIdentifierPerFile[fileName].find((moduleNameIdentifier: ModuleNameIdentifier) => {
-            return moduleNameIdentifier.name === module;
-        }).identifier;
+        return this._internalModuleImportIdentifierPerFile[fileName][module];
     }
-
-    private _getMockFactoryId(declaration: ts.Declaration): string {
+    private _getMockFactoryIdForTypeofEnum(declaration: ts.EnumDeclaration): string {
         const thisFileName: string = this._fileName;
 
         if (this._factoryCache.has(declaration)) {
             return this._factoryCache.get(declaration);
         }
 
-        const key: string = this._declarationCache.get(declaration);
+        const key: string = this.getDeclarationKeyMap(declaration);
 
         this._factoryCache.set(declaration, key);
 
         this._factoryRegistrationsPerFile[thisFileName] = this._factoryRegistrationsPerFile[thisFileName] || [];
 
-        const descriptor: ts.Expression = GetDescriptor(declaration, new Scope(key));
-
-        const mockGenericParameter: ts.ParameterDeclaration = this._getMockGenericParameter();
-
-        const factory: ts.FunctionExpression = TypescriptCreator.createFunctionExpressionReturn(descriptor, [mockGenericParameter]);
+        const factory: ts.Expression = GetTypeofEnumDescriptor(declaration, new Scope(key));
 
         this._factoryRegistrationsPerFile[thisFileName].push({
             key: declaration,
@@ -227,6 +266,19 @@ export class MockDefiner {
             return this._factoryIntersectionsRegistrationsPerFile[sourceFile.fileName]
                 .map((reg: { keys: ts.Declaration[]; factory: ts.Expression }) => {
                     const key: string = this._factoryIntersectionCache.get(reg.keys);
+
+                    return this._createRegistration(sourceFile.fileName, key, reg.factory);
+                });
+        }
+
+        return [];
+    }
+
+    private _getRegisterMockInFile(sourceFile: ts.SourceFile): ts.Statement[] {
+        if (this._registerMockFactoryRegistrationsPerFile[sourceFile.fileName]) {
+            return this._registerMockFactoryRegistrationsPerFile[sourceFile.fileName]
+                .map((reg: { key: ts.Declaration; factory: ts.Expression }) => {
+                    const key: string = this._registerMockFactoryCache.get(reg.key);
 
                     return this._createRegistration(sourceFile.fileName, key, reg.factory);
                 });