+++ src/transformer/mockDefiner/modules/moduleNameIdentifier.ts
@@ -1,9 +1,6 @@
 import * as ts from 'typescript';
-import { ModuleName } from './moduleName';
-import { ModuleImportUrl } from './modulesImportUrl';
 
 export interface ModuleNameIdentifier {
-    name: ModuleName;
-    moduleUrl: ModuleImportUrl;
+    moduleUrl: string;
     identifier: ts.Identifier;
 }